WIP (Move)

+++ src/main.go
@@ -129,6 +129,7 @@ func main() {
 	app.Get("/file", file)
 	app.Get("/fsOps/del", fsDel)
 	app.Get("/fsOps/rename", fsRename)
+	app.Get("/fsOps/move", fsMove)
 
 	subFS, _ := fs.Sub(static, "static")
 	app.Use("/", filesystem.New(filesystem.Config{
@@ -293,6 +294,14 @@ func file(c *fiber.Ctx) error {
 	return c.SendFile(fullPath)
 }
 
+func fileExists(filename string) bool {
+	info, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+	return !info.IsDir()
+}
+
 func fsDel(c *fiber.Ctx) error {
 	if c.Locals("readOnly").(bool) {
 		return fiber.NewError(fiber.StatusForbidden, "Read-only mode enabled")
@@ -336,6 +345,45 @@ func fsRename(c *fiber.Ctx) error {
 
 	fullPath := filepath.Join(root, path)
 	newPath := filepath.Join(filepath.Dir(fullPath), nuName)
+
+	if fileExists(newPath) {
+		return fiber.NewError(fiber.StatusBadRequest, "File already exists")
+	}
+
+	if err := os.Rename(fullPath, newPath); err != nil {
+		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
+	}
+
+	return c.SendStatus(200)
+}
+
+func fsMove(c *fiber.Ctx) error {
+	if c.Locals("readOnly").(bool) {
+		return fiber.NewError(fiber.StatusForbidden, "Read-only mode enabled")
+	}
+
+	if err := doAuth(c, c.Locals("pwdHash").(string)); err != nil {
+		return err
+	}
+
+	path := c.Query("path")
+	destDir := c.Query("destDir")
+	if path == "" || destDir == "" {
+		return fiber.ErrNotFound
+	}
+
+	root := c.Locals("root").(string)
+
+	fullPath := filepath.Join(root, path)
+	newPath := filepath.Join(root, destDir, filepath.Base(fullPath))
+
+	println(fullPath)
+	println(newPath)
+
+	if fileExists(newPath) {
+		return fiber.NewError(fiber.StatusBadRequest, "File already exists")
+	}
+
 	if err := os.Rename(fullPath, newPath); err != nil {
 		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 	}
